Code refactoring and addressing the review comments
