+++ lib/pbench/client/__init__.py
@@ -38,8 +38,8 @@ class API(Enum):
     name completion.
     """
 
-    COMPARE_DATASETS = "compare_datasets"
     DATASETS = "datasets"
+    DATASETS_COMPARE = "datasets_compare"
     DATASETS_CONTENTS = "datasets_contents"
     DATASETS_DETAIL = "datasets_detail"
     DATASETS_INVENTORY = "datasets_inventory"
