+++ lib/pbench/test/unit/server/test_endpoint_configure.py
@@ -55,11 +55,11 @@ class TestEndpointConfig:
         expected_results = {
             "identification": f"Pbench server {server_config.COMMIT_ID}",
             "uri": {
-                "compare_datasets": {"template": f"{uri}/compare", "params": {}},
                 "datasets": {
                     "template": f"{uri}/datasets/{{dataset}}",
                     "params": {"dataset": {"type": "string"}},
                 },
+                "datasets_compare": {"template": f"{uri}/compare", "params": {}},
                 "datasets_contents": {
                     "template": f"{uri}/datasets/{{dataset}}/contents/{{target}}",
                     "params": {
